chore(insights): extract EditorFiltersShell with CSSTransition, MaxTool, and banners
Moves the CSSTransition wrapper, SessionAnalysisWarning, MaxTool
integration, and SuggestionBanner into EditorFiltersShell. EditorFilters
now just passes filter content as children.
